+++ Mitosis.Tests/HashSetCloneEngineTests.cs
@@ -17,4 +17,17 @@ public sealed class HashSetCloneEngineTests
 		Assert.AreNotSame(obj, copy);
 		Assert.IsTrue(obj.SetEquals(copy));
 	}
+	
+	[Test]
+	public void TestIntsWithInterface()
+	{
+		var valueEngine = new DefaultCloneEngine();
+		var engine = new HashSetCloneEngine(valueEngine);
+		
+		ISet<int> obj = new HashSet<int> { 1, 2, 3 };
+		var copy = engine.Clone(obj);
+
+		Assert.AreNotSame(obj, copy);
+		Assert.IsTrue(obj.SetEquals(copy));
+	}
 }